requirements.txt
fastapi>=0.111,<1
uvicorn[standard]>=0.24
pydantic>=2.6
pydantic-settings>=2.2
python-multipart>=0.0.9
google-cloud-storage>=2.16
PyYAML>=6.0
SQLAlchemy>=2.0
